refactor: separate dag set scheduler
